Fix CAS redirects not working
The http.Request.URL field on the server side is not a fully qualfied URL and is only parsed from the Request-URI header. As such the redirect to CAS was sending an unqualfied service parameter and additionally attempting ticket validation with an unqualfied service parameter.

New requestURL function has been added to build the URL from the http.Request properties.